requirements.txt
flask==3.1.2
numpy==1.24.3
pandas==2.1.0
scikit-learn==1.8.0
matplotlib==3.8.0
seaborn==0.12.2
python-dotenv==1.0.0